Refactored some shit
