+++ pages/book/index.js
@@ -1,5 +1,5 @@
 import {fetchClient} from "../../utils.js";
-import {createStar} from "./BookElementsFactory.js";
+import {createStar} from "./bookReviewFactory.js";
 import * as Books from "./BookLists/UserBooks.js"
 
 const bookReference = "_ojXNuzgHRcC"
@@ -7,9 +7,9 @@ const bookReference = "_ojXNuzgHRcC"
 export const initBook = (referenceId) => {
   Books.init(referenceId)
       .then(() => {
-        setupFavButton()
+        setupFavButton(referenceId)
         initReviewStars(3)
-        updatePageDetails(Books.getBook(),Books.bookListTitles())
+        updatePageDetails(Books.getBook(),Books.getBookLists())
       })
 };
 
@@ -26,17 +26,22 @@ const updatePageDetails = (book,bookLists) => {
   posterElement.style.backgroundSize = "cover"
   bookLists.forEach((b,i) => {
     const opt = document.createElement("option")
-    opt.textContent = b
+    opt.textContent = b.title
     if(i === 0)
       opt.selected = true
+    opt.value = b.id
     bookListElement.appendChild(opt)
   })
   
 }
 
-const setupFavButton = () => {
+const setupFavButton = reference => {
   const btn = document.getElementById("fav-btn")
-  btn.onclick = async () => handleUpdateBookList(bookListId,bookId)
+  btn.onclick = async () => {
+    const bookListSelector = document.getElementById("list-sel")
+    const bookListReference = bookListSelector.value
+    Books.updateBookList(reference,bookListReference)
+  }
 }
 
 const handleUpdateBookList = async (bookListId, bookId) => {
